Add first pass at a DST change reminder
