+++ package.json
@@ -4,12 +4,14 @@
   "description": "Posts Slack notifications from ProjectWorks",
   "main": "handler.js",
   "scripts": {
-    "lint": "eslint . --ext js,jsx,ts,tsx",
     "deploy-staging": "serverless deploy --stage=staging",
     "deploy-prod": "serverless deploy --stage=prod",
     "coverage": "jest --silent --colors --coverage",
-    "lint-fix": "eslint . --ext js,jsx,ts,tsx --fix",
-    "test": "jest --colors --silent --passWithNoTests"
+    "test": "jest --colors --silent --passWithNoTests",
+    "format-check": "prettier --check './**/*.{css,scss,json,md,js,ts,tsx,jsx}'",
+    "format-fix": "prettier --write './**/*.{css,scss,json,md,js,ts,tsx,jsx}'",
+    "js-lint": "eslint . --ignore-pattern '!.eslintrc.js' --ext js,ts,tsx,jsx",
+    "js-lint-fix": "eslint . --ignore-pattern '!.eslintrc.js' --ext js,ts,tsx,jsx --fix"
   },
   "engines": {
     "node": ">=16.0.0"
@@ -24,6 +26,7 @@
     "@slack/webhook": "^5.0.3",
     "date-fns": "^2.16.1",
     "date-fns-tz": "^1.0.12",
+    "luxon": "^3.0.3",
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
